+++ fri/__init__.py
@@ -8,8 +8,8 @@ from fri.genData import genRegressionData, genClassificationData, genOrdinalRegr
 from fri.main import FRIBase
 from fri.model.classification import Classification
 from fri.model.lupi_classification import LUPI_Classification
-from fri.model.lupi_ordinal_regression import LUPI_OrdinalRegression
-from fri.model.lupi_ordinal_regression_exp import LUPI_OrdinalRegression_EXP
+from fri.model.lupi_ordinal_regression_imp import LUPI_OrdinalRegression_IMP
+from fri.model.lupi_ordinal_regression_exp import LUPI_OrdinalRegression
 from fri.model.lupi_regression import LUPI_Regression
 from fri.model.ordinal_regression import OrdinalRegression
 from fri.model.ordinal_regression_imp import OrdinalRegression_Imp
@@ -25,7 +25,7 @@ class ProblemName(Enum):
     LUPI_CLASSIFICATION = LUPI_Classification
     LUPI_REGRESSION = LUPI_Regression
     LUPI_ORDREGRESSION = LUPI_OrdinalRegression
-    LUPI_ORDREGRESSION_EXP = LUPI_OrdinalRegression_EXP
+    LUPI_ORDREGRESSION_IMP = LUPI_OrdinalRegression_IMP
 
 
 __all__ = ["genRegressionData", "genClassificationData", "genOrdinalRegressionData", "quick_generate",
